server/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
decouple==0.0.7
django==4.1.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycparser==2.21
pytest==7.1.3
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
psycopg2-binary==2.9.4
django-ratelimit==3.0.1
openpyxl~=3.0.10
requests~=2.28.2